backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
python-jose[cryptography]
httpx
python-multipart
python-dotenv